Refine services description in Info.plist
This splits the service offerings in two: One for files, one for text selections. The use of NSSendFileTypes limits the service to only showing up in Finder for supported filetypes. The text selection service description is changed to match the description TextEdit uses for its "Open selection" service. Note that the file service is only compatible with 10.6+.