Fix exception when using Mountain Lion's press-and-hold to insert a special character
Specifically, when you hit a number key (from the pop-up menu) to select which special character. I suspect this causes a nil character to be inserted, or something like that; regardless, and an exception from trying to call a non-existent isEqualToString on it was the result. Switching around the literals and the parameter is a MUCH safer approach, and fixes this exception.