merge dbFixes branch
with the "kUseCachesFolderForInterimNoteChanges" compile flag set to 1
nvALT will now store its "Interim Note-Changes" file in
'~/Library/Caches/net.elasticthreads.nv/'

there's also a fix in here about how nvALT's window activates on launch
when the Dock icon is hiding. Though innocous, it's kind of the straw
that broke the full 10.6 support. From now on we'll have to release a
separate 10.6 build. It requires two small changes to info.plist (see
below)

info.plist is set for 10.7+ only
to build for 10.6 make the following changes to info.plist:
- change "Minimum System Version" to 10.6.8
- change "Application is agent (UIElement)"   (aka LSUIElement) to true